Get rid of InvertibleCommandClause, call as part of build/validation flow
