+++ src/main/java/io/stargate/sgv2/jsonapi/api/model/command/clause/filter/FilterOperator.java
@@ -1,6 +1,9 @@
 package io.stargate.sgv2.jsonapi.api.model.command.clause.filter;
 
-/** This interface will be implemented by Operator enum. */
+/**
+ * This interface will be implemented by various comparison operator enums (like {@link
+ * ArrayComparisonOperator}).
+ */
 public interface FilterOperator {
   String getOperator();
